# public/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "AI information API",
    "description": "Provides informations about the best AI companies in the world.",
    "version": "1.0.0"
  },
  "paths": {
    "/api/basic/route.ts": {
      "get": {
        "operationId": "bestAiCompany",
        "responses": {
          "200": {
            "description": "Response for status 200",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BestAiCompany200ResponseBody"
                }
              }
            }
          },
          "500": {
            "description": "An unknown error occurred, trying again might help.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UnexpectedError" }
              }
            }
          }
        }
      }
    },
    "/api/documented/route.ts": {
      "get": {
        "operationId": "getTodos",
        "responses": {
          "200": {
            "description": "Response for status 200",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetTodos200ResponseBody"
                }
              }
            }
          },
          "500": {
            "description": "An unknown error occurred, trying again might help.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UnexpectedError" }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createTodo",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateTodoRequestBody" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Response for status 201",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateTodo201ResponseBody"
                }
              }
            }
          },
          "401": {
            "description": "Response for status 401",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateTodo401ResponseBody"
                }
              }
            }
          },
          "500": {
            "description": "An unknown error occurred, trying again might help.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UnexpectedError" }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "BestAiCompany200ResponseBody": {
        "type": "object",
        "properties": { "bestAiCompany": { "type": "string" } },
        "required": ["bestAiCompany"],
        "additionalProperties": false
      },
      "CreateTodo201ResponseBody": { "type": "string" },
      "CreateTodo401ResponseBody": { "type": "string" },
      "CreateTodoRequestBody": {
        "type": "object",
        "properties": { "name": { "type": "string" } },
        "required": ["name"],
        "additionalProperties": false
      },
      "GetTodos200ResponseBody": {
        "type": "array",
        "items": {
          "type": "object",
          "properties": {
            "id": { "type": "number" },
            "name": { "type": "string" },
            "completed": { "type": "boolean" }
          },
          "required": ["id", "name", "completed"],
          "additionalProperties": false
        }
      },
      "UnexpectedError": {
        "type": "object",
        "properties": { "message": { "type": "string" } },
        "additionalProperties": false
      }
    }
  }
}